Added MVC Architechture
Created widgets folder for Controller. Minimized line of code in individual files. Removed wish from providers list and kept only wishlist model as a provider.
