+++ lib/screens/home.dart
@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:wish_pool/models/wishlist.dart';
 import 'package:wish_pool/screens/add_wish.dart';
-
-import 'edit_wish.dart';
-import 'package:provider/provider.dart';
+import 'package:wish_pool/widgets/wishlist_container.dart';
 
 class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key? key}) : super(key: key);
   static const routeName = '/home';
+  const HomeScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -15,62 +12,7 @@ class HomeScreen extends StatelessWidget {
       appBar: AppBar(
         title: const Text("Wish Pool"),
       ),
-      body: Padding(
-        padding: const EdgeInsets.only(
-          top: 40,
-          left: 20,
-          right: 20,
-        ),
-        child: Consumer<WishList>(
-          builder: (context, wishes, child) => ListView.builder(
-            itemCount: wishes.totalWishes,
-            itemBuilder: ((context, index) => Card(
-                  child: Padding(
-                    padding: const EdgeInsets.all(10.0),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(wishes.allWishes[index].title),
-                            Text(wishes.allWishes[index].description),
-                          ],
-                        ),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          children: [
-                            IconButton(
-                              onPressed: () {
-                                Navigator.pushNamed(
-                                  context,
-                                  EditWish.routeName,
-                                  arguments: wishes.allWishes[index],
-                                );
-                              },
-                              icon: const Icon(
-                                Icons.edit,
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 2,
-                            ),
-                            IconButton(
-                              icon: const Icon(
-                                Icons.delete,
-                              ),
-                              onPressed: () =>
-                                  wishes.removeWish(wishes.allWishes[index]),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                )),
-          ),
-        ),
-      ),
+      body: const WishlistContainer(),
       floatingActionButton: IconButton(
         icon: const Icon(Icons.add),
         onPressed: () {
